+++ CHANGELOG.md
@@ -301,6 +301,7 @@ _(none)_
 * Fix `error404document` mapping on `azure.storage.Account`
 * Upgrade to v2.9.0 of the AzureRM Terraform Provider
 * Update ArchiveFunctionApp to pass correct params to FunctionApp
+* Allow users to specify their own http output binding for HTTP triggered Azure Functions
 
 ## 3.4.0 (2020-05-01)
 * Upgrade to v2.8.0 of the AzureRM Terraform Provider